Organize relative files by concern
It's a commom pratic in react projects separate relative
files by folders to help to scalate the project.

See more:

* https://medium.com/@alexmngn/how-to-better-organize-your-react-applications-2fd3ea1920f1
