+++ src/components/App/index.js
@@ -1,18 +1,18 @@
 import React from 'react';
-import './App.scss';
+import './index.scss';
 
-import Members from './components/Members';
-import Events from './components/Events';
-import FixedNavbar from './components/FixedNavbar';
+import Members from '../Members';
+import Events from '../Events';
+import FixedNavbar from '../FixedNavbar';
 
-import letras_epic from './img/letras_epic.png';
-import background_icpc from './img/background_icpc.jpg';
-import logo_epic from './img/bolt_epic.png';
-import logo_icpc from './img/logo_icpc.png';
-import logo_mdp from './img/logo_mdp.jpg';
-import logo_mineira from './img/logo_mineira.png';
-import logo_b2ml from './img/logo_b2ml.png';
-import logo_unifei from './img/logo_unifei.png';
+import letras_epic from '../../img/letras_epic.png';
+import background_icpc from '../../img/background_icpc.jpg';
+import logo_epic from '../../img/bolt_epic.png';
+import logo_icpc from '../../img/logo_icpc.png';
+import logo_mdp from '../../img/logo_mdp.jpg';
+import logo_mineira from '../../img/logo_mineira.png';
+import logo_b2ml from '../../img/logo_b2ml.png';
+import logo_unifei from '../../img/logo_unifei.png';
 
 let about = "A competição de programação, é uma prova de 5 horas, com 8 a 14 problemas para serem resolvidos. Existem 4 competições de programação que a UNIFEI vem participando. A principal é a Maratona de Programação. Ela é dividida em 3 fases no Brasil. A primeira é uma competição regional, realizada sempre em setembro de cada ano, em uma sede em cidade próxima a Itajubá. Essa etapa, através das regras criadas pela organização, classifica para a fase final brasileira ou regional da América Latina. As equipes que passarem na primeira fase, geralmente em torno de 60 a 70 equipes do Brasil inteiro, se reúnem em novembro em uma cidade pré-definida para disputarem uma das 5 ou 6 vagas para a etapa Mundial. O Mundial é em um país e cidade pré-definida e ocorre em maio do ano seguinte.";
