+++ src/components/App/index.scss
