+++ src/components/Events/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import data from './data.json';
+import data from '../../data.json';
 
 class Event extends React.Component {
